Final Year Project
This is the repository for my final year university project at the University of Kent. The aim of the project is to create a recipea app that allows users to search by ingredient. This was build with the Ionic Framework V2, please see there docs for any details http://ionicframework.com/docs/